First implementation of backend
