--- Signup.html
@@ -1,134 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
-        crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css"
-        integrity="sha512-z3gLpd7yknf1YoNbCzqRKc4qyor8gaKU1qmn+CShxbuBusANI9QpRohGBreCFkKxLhei6S9CQXFEbbKuqLg0DA=="
-        crossorigin="anonymous" referrerpolicy="no-referrer" />
-    <link rel="stylesheet" href="Login.css">
-    <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <title>Sign Up</title>
-</head>
-
-<body>
-    <!-- Navbar -->
-    <nav class="navbar navbar-expand-lg bg-body-tertiary">
-        <div class="container">
-            <!-- Logo and Website Name -->
-            <a class="navbar-brand fs-4" href="#">Woolly Wonders</a>
-            <!-- Toggle Menu -->
-            <button class="navbar-toggler shadow-none border-bottom-0" type="button" data-bs-toggle="offcanvas"
-                data-bs-target="#offcanvasNavbar" aria-controls="offcanvasNavbar" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <!-- Side Menu Bar -->
-            <div class="sidebar offcanvas offcanvas-end" tabindex="-1" id="offcanvasNavbar"
-                aria-labelledby="offcanvasNavbarLabel">
-                <!-- Side Menu Header -->
-                <div class="offcanvas-header text-white border-bottom">
-                    <h5 class="offcanvas-title" id="offcanvasNavbarLabel">Woolly Wonders</h5>
-                    <button type="button" class="btn-close btn-close-white shadow-none" data-bs-dismiss="offcanvas"
-                        aria-label="Close"></button>
-                </div>
-                <!-- Side Menu Body -->
-                <div class="offcanvas-body d-flex flex-column p-4 flex-lg-row p-lg-0">
-                    <ul class="navbar-nav justify-content-center align-items-center flex-grow-1 fs-5 pe-3">
-                        <li class="nav-item mx-2">
-                            <a class="nav-link" aria-current="page" href="#home">Home</a>
-                        </li>
-                        <li class="nav-item mx-2">
-                            <a class="nav-link" href="#about">About</a>
-                        </li>
-                        <li class="nav-item mx-2">
-                            <a class="nav-link" href="#services">Services</a>
-                        </li>
-                        <li class="nav-item mx-2">
-                            <a class="nav-link" href="#contact">Contact</a>
-                        </li>
-                    </ul>
-                    <!-- Login And Sign Up -->
-                    <div class="d-flex flex-column justify-content-center align-items-center gap-3 flex-lg-row">
-                        <a href="#login" class="login-btn text-decoration-none px-3 ">Login</a>
-                        <a href="#signup" class="sign-up-btn text-white text-decoration-none px-3 py-1 rounded-4">Sign
-                            Up</a>
-                    </div>
-                    <!-- Thai' WILL USE THIS LATTER PLEASE DO NOT DELETE -->
-                    <!-- <form class="d-flex mt-3" role="search">
-                        <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-                        <button class="btn btn-outline-success" type="submit">Search</button>
-                    </form> -->
-                </div>
-            </div>
-        </div>
-    </nav>
-
-    <main>
-        <section class="container-login forms">
-            <!-- Sign Up --> 
-            <div class="form login">
-                <div class="form-content">
-                    <header>Sign Up</header>
-                    <form action="#">
-                        <div class="field input-field">
-                            <input type="email" placeholder="Email" class="email">
-                        </div>
-                        <div class="field input-field">
-                            <input type="password" placeholder="Password" class="password">
-                            <i class='bx bx-hide hide-icon'></i>
-                        </div>
-                        <div class="field input-field">
-                            <input type="password" placeholder=" Re-enter Password" class="password">
-                            <i class='bx bx-hide hide-icon'></i>
-                        </div>
-                        <div class="field button">
-                            <button>Sign Up</button>
-                        </div>
-                        <div class="form-link">
-                            <span>Already have an account?<a href="Login.html" class="link login-link"> Login</a></span>
-                        </div>
-                    </form>                  
-                </div>
-            </div>
-        </section>
-    </main>
-
-    <footer class="footer border-top">
-        <div class="container-footer grid--footer">
-            <div class="logo-col">
-                <!-- Logo Here -->
-                <a class="footer-logo" href="#">
-                    <img class="logo" src="" alt="website-logo">
-                </a>
-                <p class="copyright">Copyright &copy; 2023 by Thai Nguyen. All right reserved</p>
-                <p class="copyright">
-                    Handi Resco Tower, 521 Kim Ma, Ba Dinh, Ha Noi
-                </p>
-                <ul class="social-links">
-                    <li><a href="#"><i class="social-icon fa-brands fa-square-facebook"></i></a></li>
-                    <li><a href="#"><i class="social-icon fa-brands fa-instagram"></i></a></li>
-                    <li><a href="#"><i class="social-icon fa-brands fa-github"></i></a></li>
-                </ul>
-            </div>
-            <nav class="nav-col">
-                <p class="footer-heading">More Information</p>
-                <ul class="footer-nav">
-                    <li><a class="footer-link" href="#">About Us</a></li>
-                    <li><a class="footer-link" href="#">Copyright</a></li>
-                    <li><a class="footer-link" href="#">Privacy & terms</a></li>
-                    <li><a class="footer-link" href="#">Help Links</a></li>
-                </ul>
-            </nav>
-        </div>
-    </footer>
-    <!-- JavaScript -->
-    <script src="Signin.js"></script>
-</body>
-
-</html>
