+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "asm2_web",
+  "version": "1.0.0",
+  "description": "COSC2430 - Assignment 2",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "body-parser": "^1.20.2",
+    "ejs": "^3.1.9",
+    "express": "^4.18.2",
+    "mongoose": "^7.5.1"
+  }
+}
