+++ public/styles/Login.css
@@ -1,0 +1,225 @@
+:root {
+    --component-color: #ff385c
+}
+
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+.container-login{
+    height: 70vh;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.form{
+    border: 1px solid rgba(0,0,0, 0.7);
+    border-radius: 6px;
+    max-width: 430px;
+    width: 100%;
+    padding: 30px;
+    box-shadow: 0px 1px 6px 1px rgba(22, 18, 18, 0.7);
+}
+
+header{
+    font-size: 28px;
+    font-weight: 600;
+    text-align: center;
+}
+
+form{
+    margin-top: 30px;
+}
+
+form .field{
+    position: relative;
+    height: 50px;
+    width: 100%;
+    margin-top: 20px;
+}
+
+.field input, .field button{
+    height: 100%;   
+    width: 100%;
+    border: none;
+    border-radius: 6px;
+    font-size: 16px;
+    font-weight: 400;
+}
+
+.field input{
+    border: 1px solid #cacaca;
+    padding: 0 15px;    
+}
+
+.hide-icon{
+    position: absolute;
+    top: 35%;
+    right: 10px;
+    color: #8b8b8b;
+    font-size: 18px;
+    cursor: pointer;
+}
+
+.field button{
+    background-color: var(--component-color);
+    color: white;
+    transition: 0.2s ease-out;
+}
+
+.field button:active{
+    transform: scale(0.96);
+    background-color: #e7193f;
+}
+
+.form-container{
+    display: flex;
+    justify-content: space-between;
+    flex-direction: row;
+    margin-top: 10px;
+}
+
+.form-link {
+    margin-top: 10px;
+    text-align: center;
+}
+
+.form-link span, form link a{
+    font-size: 14px;
+    font-weight: 400;
+}
+
+.form a{
+    text-decoration: none;
+}
+
+.form a:hover{
+    text-decoration: none;
+    color: blue;
+}
+
+@media (max-width: 992px) {
+    .sidebar {
+        /* background-color: rgba(255, 255, 255, 0.15); */
+        background: linear-gradient(0.85turn ,#fff, #e6afb9);
+    }
+}
+
+.nav-link {
+    color: #222;
+    border-bottom: 1px solid transparent;
+}
+
+.nav-link:hover,
+.nav-link:active {
+    border-bottom: 1px solid var(--component-color);
+    color: var(--component-color);
+}
+
+.login-btn {
+    color: var(--component-color);
+}
+
+.sign-up-btn {
+    background-color: var(--component-color);
+}
+
+/* footer section */
+
+.grid--footer {
+    display: grid;
+    column-gap: 3.2rem;
+    gap: 3.2rem;
+    align-items: center;
+    justify-content: center;
+    grid-template-columns: 1fr 1fr;
+}
+
+.footer {
+    padding: 3.2rem 9.6rem;
+    /* border-top: 1px solid #CACACA; */
+    background-color: #fff;
+}
+
+.logo-col {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.footer-logo {
+    display: block;
+    margin-bottom: 2.4rem;
+}
+
+.social-links {
+    list-style: none;
+    display: flex;
+    justify-content: center;
+    gap: 2.4rem;
+    padding: 0;
+}
+
+.nav-col {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.social-icon {
+    margin-top: 1rem;
+    font-size: 1.8rem;
+    color: #222;
+}
+
+.social-icon:hover,
+.social-icon:active {
+    transition: all 0.6s;
+    transform: translateY(-20%);
+    color: var(--component-color);
+}
+
+.copyright {
+    font-size: 1rem;
+    line-height: 1.6;
+    color: #222;
+    margin-top: auto;
+}
+
+.footer-heading {
+    font-size: 1.2rem;
+    font-weight: 500;
+    margin-bottom: 2rem;
+}
+
+.footer-nav {
+    list-style: none;
+    display: flex;
+    flex-direction: column;
+    gap: 1.2rem;
+}
+
+.footer-link:link,
+.footer-link:visited {
+    text-decoration: none;
+    font-size: 1rem;
+    color: #222;
+    transition: all 0.3s;
+}
+
+.footer-link:hover,
+.footer-link:active {
+    color: var(--component-color);
+}
+
+@media (max-width:768px){
+    .form{
+        padding: 15px 10px;
+        border: none;
+        box-shadow: none;
+    }
+}
