+++ views/Login.ejs
@@ -1,0 +1,42 @@
+<!-- Adding EJS header -->
+<%- include("partials/header.ejs") %>
+
+<!-- HTML code for login -->
+
+<section class="container-login forms">
+    <!-- Log in -->
+    <div class="form login">
+        <div class="form-content">
+            <header>Login</header>
+            <form action="#">
+                <div class="field input-field">
+                    <input type="email" placeholder="Email" class="email">
+                </div>
+                <div class="field input-field">
+                    <input type="password" placeholder="Password" class="password">
+                    <i class='bx bx-hide hide-icon'></i>
+                </div>
+                <div class="form-container">
+                    <div class="form-check">
+                        <input class="form-check-input" type="checkbox" value="" id="remember-pass">
+                        <label class="form-check-label" for="remember-pass">
+                            Remember Password
+                        </label>
+                    </div>
+                    <div class="form-link-pass">
+                        <a href="#" class="forgot-password">Forgot Password?</a>
+                    </div>
+                </div>
+                <div class="field button">
+                    <button>Login</button>
+                </div>
+                <div class="form-link">
+                    <span>Don't have an account?<a href="/signup" class="link signup-link"> Sign Up</a></span>
+                </div>
+            </form>                  
+        </div>
+    </div>
+</section>                
+   
+<!-- Adding EJS footer -->
+<%- include("partials/footer.ejs") %>
