+++ views/partials/footer.ejs
@@ -1,0 +1,35 @@
+</main>
+
+<footer class="footer border-top">
+    <div class="container-footer grid--footer">
+        <div class="logo-col">
+            <!-- Logo Here -->
+            <a class="footer-logo" href="#">
+                <img class="logo" src="" alt="website-logo">
+            </a>
+            <p class="copyright">Copyright &copy; 2023 by Thai Nguyen. All right reserved</p>
+            <p class="copyright">
+                Handi Resco Tower, 521 Kim Ma, Ba Dinh, Ha Noi
+            </p>
+            <ul class="social-links">
+                <li><a href="#"><i class="social-icon fa-brands fa-square-facebook"></i></a></li>
+                <li><a href="#"><i class="social-icon fa-brands fa-instagram"></i></a></li>
+                <li><a href="#"><i class="social-icon fa-brands fa-github"></i></a></li>
+            </ul>
+        </div>
+        <nav class="nav-col">
+            <p class="footer-heading">More Information</p>
+            <ul class="footer-nav">
+                <li><a class="footer-link" href="#">About Us</a></li>
+                <li><a class="footer-link" href="#">Copyright</a></li>
+                <li><a class="footer-link" href="#">Privacy & terms</a></li>
+                <li><a class="footer-link" href="#">Help Links</a></li>
+            </ul>
+        </nav>
+    </div>
+</footer>
+<!-- JavaScript -->
+<script src="script/Signin.js"></script>
+</body>
+
+</html>
